Increased steps for map gen. Fixed issue with weird printing.
Lose sequences and the "Exit..." message when exiting the game were not displaying correctly due to stdout not being flushed in time. Added forced-flushes to fix.

Removed whitespace character printing from the normal lose sequence as it had no effect.

Added an extra move delay to the funny dance as it felt weird without.